[Feature #152896166] Refactor project structure
