+++ src/components/ui/__tests__/ShowNotifications.js
@@ -2,7 +2,7 @@ import React from "react";
 import Enzyme, {mount} from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 import * as sinon from "sinon";
-import ShowNotifications from "../ui/ShowNotifications";
+import ShowNotifications from "../ShowNotifications";
 
 Enzyme.configure({adapter: new Adapter()});
