+++ src/store/reducers/__tests__/reducers.js
